Log tree class refactor (#3068)
* log tree class re-org

* rename

* copyrights
