+++ src/io/flutter/logging/FlutterLogColors.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 The Chromium Authors. All rights reserved.
+ * Copyright 2019 The Chromium Authors. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be
  * found in the LICENSE file.
  */
@@ -9,12 +9,12 @@ import com.intellij.ui.JBColor;
 import com.intellij.util.ui.UIUtil;
 import org.jetbrains.annotations.NotNull;
 
-import java.awt.*;
+import java.awt.Color;
 
 public class FlutterLogColors {
 
   @NotNull
-  static Color forCategory(@NotNull String category) {
+  public static Color forCategory(@NotNull String category) {
     if (category.startsWith("flutter.")) {
       return JBColor.gray;
     }
