+++ src/io/flutter/logging/FlutterLogEntryParser.java
@@ -21,9 +21,9 @@ import com.intellij.ui.SimpleTextAttributes;
 import com.intellij.util.EventDispatcher;
 import com.jetbrains.lang.dart.ide.runner.DartConsoleFilter;
 import io.flutter.console.FlutterConsoleFilter;
-import io.flutter.logging.util.LineInfo;
-import io.flutter.logging.util.LineParser;
-import io.flutter.logging.util.StyledText;
+import io.flutter.logging.text.LineInfo;
+import io.flutter.logging.text.LineParser;
+import io.flutter.logging.text.StyledText;
 import io.flutter.run.FlutterDebugProcess;
 import io.flutter.run.daemon.DaemonApi;
 import io.flutter.server.vmService.HeapMonitor;
